rework wearables
