+++ Cargo.toml
@@ -34,7 +34,7 @@ wallet = { path="crates/wallet" }
 nft = { path="crates/nft" }
 tween = { path="crates/tween" }
 world_ui = { path="crates/world_ui" }
-emotes = { path="crates/emotes" }
+collectibles = { path="crates/collectibles" }
 
 bevy = { version = "0.13", features=["jpeg"] }
 bevy_console = { git = "https://github.com/robtfm/bevy-console", branch="egui-26" }
@@ -87,7 +87,7 @@ nft = { workspace = true }
 tween = { workspace = true }
 scene_material = { workspace = true }
 world_ui = { workspace = true }
-emotes = { workspace = true }
+collectibles = { workspace = true }
 
 bevy = { workspace = true }
 bevy_console = { workspace = true }
