+++ crates/avatar/Cargo.toml
@@ -17,7 +17,7 @@ ui_core = { workspace = true }
 wallet = { workspace = true }
 input_manager = { workspace = true }
 world_ui = { workspace = true }
-emotes = { workspace = true }
+collectibles = { workspace = true }
 scene_material = { workspace = true }
 
 bevy = { workspace = true }
