+++ crates/system_ui/Cargo.toml
@@ -17,7 +17,7 @@ avatar = { workspace = true }
 input_manager = { workspace = true }
 av = { workspace = true }
 wallet = { workspace = true }
-emotes = { workspace = true }
+collectibles = { workspace = true }
 tween = { workspace = true }
 console = { workspace = true }
 scene_material = { workspace = true }
