+++ src/main.rs
@@ -21,6 +21,7 @@ use bevy::{
 };
 use bevy_console::ConsoleCommand;
 
+use collectibles::CollectiblesPlugin;
 use common::{
     sets::SetupSets,
     structs::{
@@ -29,7 +30,6 @@ use common::{
     },
     util::UtilsPlugin,
 };
-use emotes::EmotesPlugin;
 use restricted_actions::RestrictedActionsPlugin;
 use scene_material::SceneBoundPlugin;
 use scene_runner::{
@@ -235,7 +235,7 @@ fn main() {
         .add_plugins(NftShapePlugin)
         .add_plugins(TweenPlugin)
         .add_plugins(SceneBoundPlugin)
-        .add_plugins(EmotesPlugin)
+        .add_plugins(CollectiblesPlugin)
         .add_plugins(WorldUiPlugin);
 
     if !no_avatar {
